% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigTen_palettes.R
\docType{data}
\name{bigten_colors_primary}
\alias{bigten_colors_primary}
\title{Named vector of Big Ten primary colors for reliable institution matching}
\format{
An object of class \code{character} of length 22.
}
\usage{
bigten_colors_primary
}
\description{
Named vector of Big Ten primary colors for reliable institution matching
}
\keyword{datasets}
