% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_black}
\alias{msu_black}
\title{MSU black hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_black
}
\description{
MSU black hex colour
}
\keyword{datasets}
