% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_blue}
\alias{msu_blue}
\title{MSU blue (aka blue-grey or blue-gray) hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_blue
}
\description{
MSU blue (aka blue-grey or blue-gray) hex colour
}
\keyword{datasets}
