% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_darkgrey}
\alias{msu_darkgrey}
\title{MSU dark grey hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_darkgrey
}
\description{
MSU dark grey hex colour
}
\keyword{datasets}
