% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_Excellence}
\alias{msu_Excellence}
\title{MSU Excellence Green hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_Excellence
}
\description{
MSU Excellence Green hex colour
}
\keyword{datasets}
