% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_green}
\alias{msu_green}
\title{MSU Spartan Green hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_green
}
\description{
MSU Spartan Green hex colour
}
\keyword{datasets}
