% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_grey}
\alias{msu_grey}
\title{MSU grey/gray hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_grey
}
\description{
MSU grey/gray hex colour
}
\keyword{datasets}
