% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_Kelly}
\alias{msu_Kelly}
\title{MSU Kelly Green hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_Kelly
}
\description{
MSU Kelly Green hex colour
}
\keyword{datasets}
