% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_Lime}
\alias{msu_Lime}
\title{MSU Lime Green hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_Lime
}
\description{
MSU Lime Green hex colour
}
\keyword{datasets}
