% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_orange}
\alias{msu_orange}
\title{MSU orange hex colour}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
msu_orange

msu_orange
}
\description{
MSU orange hex colour

MSU (burnt) orange hex colour
}
\keyword{datasets}
