% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_peach}
\alias{msu_peach}
\title{MSU peach hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_peach
}
\description{
MSU peach hex colour
}
\keyword{datasets}
