% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_purple}
\alias{msu_purple}
\title{MSU purple hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_purple
}
\description{
MSU purple hex colour
}
\keyword{datasets}
