% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_Refresh}
\alias{msu_Refresh}
\title{MSU Refresh Green hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_Refresh
}
\description{
MSU Refresh Green hex colour
}
\keyword{datasets}
