% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_splitpea}
\alias{msu_splitpea}
\title{MSU split pea hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_splitpea
}
\description{
MSU split pea hex colour
}
\keyword{datasets}
