% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_teal}
\alias{msu_teal}
\title{MSU teal hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_teal
}
\description{
MSU teal hex colour
}
\keyword{datasets}
