% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_white}
\alias{msu_white}
\title{MSU white hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_white
}
\description{
MSU white hex colour
}
\keyword{datasets}
