% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{msu_yellow}
\alias{msu_yellow}
\title{MSU yellow (aka yellow-green or grellow) hex colour}
\format{
An object of class \code{character} of length 1.
}
\usage{
msu_yellow
}
\description{
MSU yellow (aka yellow-green or grellow) hex colour
}
\keyword{datasets}
