% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSU_palettes.R
\docType{data}
\name{MSUpalettes}
\alias{MSUpalettes}
\title{MSU palettes and MSU color variables}
\format{
An object of class \code{list} of length 11.
}
\usage{
MSUpalettes
}
\description{
This file contains only MSU palettes and MSU colour variables.
}
\keyword{datasets}
