% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_MSU_palette.R
\name{set_msu_palette}
\alias{set_msu_palette}
\title{Set MSU base R plotting palette}
\usage{
set_msu_palette(palette)
}
\arguments{
\item{palette}{Name of palette. See \code{names(MSUthemes::MSUpalettes)}.}
}
\value{
Returns a character vector giving the colors from the palette which
was in effect. This is invisible unless the argument is omitted.
}
\description{
Set MSU base R plotting palette
}
\examples{
set_msu_palette("msu_qual1")
}
