% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_MSU_par.R
\name{set_msu_par}
\alias{set_msu_par}
\title{Set MSUthemes base R graphical parameters}
\usage{
set_msu_par(
  family = "Metropolis",
  adj = 0,
  mar = c(5, 3, 3, 2.5),
  bty = "n",
  ...
)
}
\arguments{
\item{family}{Font used for all text elements. Default "Metropolis".}

\item{adj}{Alignment of text for title. Default 0.}

\item{mar}{Margins. Default \code{c(5, 3, 3, 2.5)}.}

\item{bty}{Axis lines. Default "n".}

\item{...}{Additional arguments passed to \code{par}}
}
\value{
Returns an invisible named list.
}
\description{
Set MSUthemes base R graphical parameters
}
\examples{
# save user's current par values that this function will change
oldpar <- par("family", "adj", "mar", "bty")
set_msu_par()
plot(1:4, 1:4, col=1:4, main = "Title")
par(oldpar)
}
